requirements.txt
gTTS==2.5.1
keras==2.10.0
mediapipe==0.10.11
numpy==1.26.4
opencv_contrib_python==4.9.0.80
pandas==2.2.2
pygame==2.5.2
tensorflow==2.10.1
Flask==3.0.2
tables==3.9.2
PyQt5==5.15.9
scikit-learn==1.7.0
joblib==1.5.1
threadpoolctl==3.6.0
mysql-connector-python==8.4.0
